refactor: move events to CorruptedVaultHelper
